Compilation steps refactor and adding inline worker (#65)
Changing create_db_from_results to be both normalized db with relational data, and make it so the assocs table only gets data it doesn't already have.

Changing the compile_results step to simplify the logic and move it to the db.

Adding worker that can ingest a new GWAS in the pipeline